requirements.txt
flask
flask_cors
flask_mysqldb
flask_security_too
flask_sqlalchemy
flask_jwt_extended
Flask-Bcrypt
flask_compress
flask_assets
pillow
pandas
datetime
openpyxl
python-dateutil
reportlab
cryptography
num2words
waitress
flask_pjax